% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpaintr.R
\name{aggregateYhat}
\alias{aggregateYhat}
\title{Aggregate predicted leave-one-out probabilities over meta variables}
\usage{
aggregateYhat(sce, target, group, assay_type = "tfmfeatures")
}
\arguments{
\item{sce}{\code{\link[SingleCellExperiment]{SingleCellExperiment}} object}

\item{target}{Name of target variable for prediction}

\item{group}{Grouping variable for cross-validation, e.g., patient}

\item{assay_type}{A string specifying the assay}
}
\value{
\code{data.frame}
}
\description{
Aggregate predicted leave-one-out probabilities over meta variables
}
