% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpaintr.R
\name{combinePanels}
\alias{combinePanels}
\title{Combine predicted leave-one-out probabilities with sample info}
\usage{
combinePanels(mae, meta_vars = c("Patient", "Treatment", "Gender"))
}
\arguments{
\item{meta_vars}{a vector of variables from `colData`}
}
\value{
\code{\link[data.frame]{data.frame}}
}
\description{
Combine predicted leave-one-out probabilities with sample info
}
