% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpaintr.R
\name{fitModel}
\alias{fitModel}
\title{Predict target from features}
\usage{
fitModel(
  sce,
  assay_type = "tfmfeatures",
  target,
  interest_level,
  reference_level,
  group,
  strata,
  n_folds = 20,
  n_threads = 32
)
}
\arguments{
\item{sce}{\code{\link[SingleCellExperiment]{SingleCellExperiment}} object}

\item{assay_type}{A string specifying the assay}

\item{target}{Name of target variable for prediction}

\item{interest_level}{Factor interest level in `target` variable}

\item{reference_level}{Factor reference level in `target` variable}

\item{group}{Grouping variable for cross-validation, e.g., patient}

\item{strata}{Stratification variable for cross-validation, e.g., joint}

\item{n_folds}{Number of cross-validation folds}

\item{n_threads}{Number of parallel threads for fitting of models}
}
\value{
\code{\link[tibble]{tibble}} data frame
}
\description{
Predict target from features
}
