% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpaintr.R
\name{generate_data}
\alias{generate_data}
\title{Simulate CellProfiler data and write to a temporary file}
\usage{
generate_data()
}
\value{
path to csv file
}
\description{
Simulate CellProfiler data and write to a temporary file
}
