% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpaintr.R
\name{loadData}
\alias{loadData}
\title{Load cell painting data from file and convert to a SingleCellExperiment}
\usage{
loadData(cell_file)
}
\arguments{
\item{cell_file}{path to csv file from CellProfiler, e.g., MyExpt_Cells.csv}
}
\value{
\code{\link[SingleCellExperiment]{SingleCellExperiment}} object
}
\description{
Load cell painting data from file and convert to a SingleCellExperiment
}
\examples{
set.seed(23)
cell_file <- generate_data()
loadData(cell_file)

}
