% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpaintr.R
\name{normalizeExclude}
\alias{normalizeExclude}
\title{Normalize by median absolute deviation and exclude zero features}
\usage{
normalizeExclude(sce)
}
\arguments{
\item{sce}{\code{\link[SingleCellExperiment]{SingleCellExperiment}} object}
}
\value{
\code{\link[SingleCellExperiment]{SingleCellExperiment}} object
}
\description{
Normalize by median absolute deviation and exclude zero features
}
