% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpaintr.R
\name{plotLOO}
\alias{plotLOO}
\title{Plot predicted leave-one-out probabilities}
\usage{
plotLOO(sce, target, group, assay_type = "tfmfeatures")
}
\arguments{
\item{sce}{\code{\link[SingleCellExperiment]{SingleCellExperiment}} object}

\item{target}{Name of target variable for prediction}

\item{group}{Grouping variable for cross-validation, e.g., patient}

\item{assay_type}{A string specifying the assay}
}
\value{
\code{\link[ggplot2]{ggplot2}} object
}
\description{
Plot predicted leave-one-out probabilities
}
\examples{
set.seed(23)
cell_file <- generate_data()
sce <- loadData(cell_file)
sce <- transformScale(sce)

sce$Drug <- as.factor(sce$Drug)
sce$Drug <- relevel(sce$Drug, ref = "D1")
types <- c("AreaShape", "Intensity", "Texture")

sce_single <- predictLOO(
    sce,
    target = "Drug", group = "Patient",
    interest_level = "D7", reference_level = "D1",
    types = types,
    n_threads = 1
)

plotLOO(sce_single, target = "Drug", group = "Patient")

}
