% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpaintr.R
\name{plotPanels}
\alias{plotPanels}
\title{Plot predicted leave-one-out probabilities}
\usage{
plotPanels(merged)
}
\arguments{
\item{merged}{a \code{\link[data.frame]{data.frame}} from `combinePanels`}
}
\value{
\code{\link[ggplot2]{ggplot2}} object
}
\description{
Plot predicted leave-one-out probabilities
}
