% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpaintr.R
\name{plotPCACor}
\alias{plotPCACor}
\title{Plot number of cells per image}
\usage{
plotPCACor(
  sce,
  filter_by = 1,
  top = 20,
  pcs = seq(5),
  assay_type = "tfmfeatures"
)
}
\arguments{
\item{sce}{\code{\link[SingleCellExperiment]{SingleCellExperiment}} object}

\item{filter_by}{PC to use for feature selection}

\item{top}{Number of top features to select}

\item{pcs}{Number of PCs to plot}

\item{assay_type}{A string specifying the assay}
}
\value{
\code{\link[ggplot2]{ggplot2}} object
}
\description{
Plot number of cells per image
}
\examples{
set.seed(23)
cell_file <- generate_data()
sce <- loadData(cell_file)
sce <- transformScale(sce, robust = TRUE)
sce <- scater::runPCA(sce, exprs_values = "tfmfeatures", ncomponents = 10)
plotPCACor(sce, filter_by = 1)

}
