% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpaintr.R
\name{plotYhat}
\alias{plotYhat}
\title{Plot predicted leave-one-out probabilities}
\usage{
plotYhat(y_hat, target = "Treatment")
}
\arguments{
\item{target}{Name of target variable for prediction}

\item{merged}{a \code{\link[data.frame]{data.frame}} from `aggregateYhat`}
}
\value{
\code{\link[ggplot2]{ggplot2}} object
}
\description{
Plot predicted leave-one-out probabilities
}
