% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpaintr.R
\name{printROC}
\alias{printROC}
\title{Print sample info of the ROC curves}
\usage{
printROC(result, fold_id)
}
\arguments{
\item{result}{\code{\link[tibble]{tibble}} data frame}

\item{fold_id}{Fold id to print}
}
\value{
NULL.
}
\description{
Print sample info of the ROC curves
}
