% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpaintr.R
\name{removeMissingValues}
\alias{removeMissingValues}
\title{Remove missing values}
\usage{
removeMissingValues(sce)
}
\arguments{
\item{sce}{\code{\link[SingleCellExperiment]{SingleCellExperiment}} object}
}
\value{
\code{\link[SingleCellExperiment]{SingleCellExperiment}} object
}
\description{
Remove missing values
}
