% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpaintr.R
\name{removeRowColEffect}
\alias{removeRowColEffect}
\title{Remove row and column effect using the median polish procedure}
\usage{
removeRowColEffect(sce, assay_type = "tfmfeatures")
}
\arguments{
\item{sce}{\code{\link[SingleCellExperiment]{SingleCellExperiment}} object}

\item{assay_type}{A string specifying the assay}
}
\value{
\code{\link[SingleCellExperiment]{SingleCellExperiment}} object
}
\description{
Remove row and column effect using the median polish procedure
}
\examples{
set.seed(23)
cell_file <- generate_data()
sce <- loadData(cell_file)
sce <- transformScale(sce)
row <- substr(sce$Well, 1, 1)
row <- factor(row, levels = LETTERS[1:length(unique(row))])
col <- substr(sce$Well, 2, 3)
col <- factor(as.numeric(col))
sce$Row <- row
sce$Col <- col
sce <- removeRowColEffect(sce)

}
