% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpaintr.R
\name{removeZeroInflation}
\alias{removeZeroInflation}
\title{Remove zero-inflated features}
\usage{
removeZeroInflation(sce, proportion = 0.2)
}
\arguments{
\item{sce}{\code{\link[SingleCellExperiment]{SingleCellExperiment}} object}

\item{proportion}{Remove features exceeding this zero-inflation proportion}
}
\value{
\code{\link[SingleCellExperiment]{SingleCellExperiment}} object
}
\description{
Remove zero-inflated features
}
