% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpaintr.R
\name{transformLogScale}
\alias{transformLogScale}
\title{Filter low variance features}
\usage{
transformLogScale(sce)
}
\arguments{
\item{sce}{\code{\link[SingleCellExperiment]{SingleCellExperiment}} object}
}
\value{
\code{\link[SingleCellExperiment]{SingleCellExperiment}} object
}
\description{
Filter low variance features
}
