% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpaintr.R
\name{transformScale}
\alias{transformScale}
\title{Transform, center, and scale features}
\usage{
transformScale(sce, robust = FALSE)
}
\arguments{
\item{sce}{\code{\link[SingleCellExperiment]{SingleCellExperiment}} object}

\item{robust}{If true robust z-score, otherwise standard z-score}
}
\value{
\code{\link[SingleCellExperiment]{SingleCellExperiment}} object
}
\description{
Transform, center, and scale features
}
\examples{
set.seed(23)
cell_file <- generate_data()
sce <- loadData(cell_file)
sce <- transformScale(sce)

}
