% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpaintr.R
\name{treatmentEffect}
\alias{treatmentEffect}
\title{Evaluate treatment effect}
\usage{
treatmentEffect(
  sce,
  assay_type = "tfmfeatures",
  group,
  treatment,
  n_perm = 100,
  n_threads = 1
)
}
\arguments{
\item{sce}{\code{\link[SingleCellExperiment]{SingleCellExperiment}} object}

\item{assay_type}{A string specifying the assay}

\item{group}{Grouping variable, e.g., patient}

\item{treatment}{Treatment variable}

\item{n_perm}{Number of permutations for the null distribution}

\item{n_threads}{Number of parallel threads for fitting of models}
}
\value{
\code{\link[ggplot2]{ggplot2}} data frame
}
\description{
Evaluate treatment effect
}
