% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpaintr.R
\name{aggregateYhat}
\alias{aggregateYhat}
\title{Aggregate predicted leave-one-out probabilities over meta variables}
\usage{
aggregateYhat(
  sce,
  assay_type = "tfmfeatures",
  meta_vars = c("Patient", "Treatment")
)
}
\arguments{
\item{sce}{\code{\link[SingleCellExperiment]{SingleCellExperiment}} object}

\item{assay_type}{A string specifying the assay}

\item{meta_vars}{a vector of variables from `colData`}
}
\value{
\code{data.frame}
}
\description{
Aggregate predicted leave-one-out probabilities over meta variables
}
\examples{
set.seed(23)
cell_file <- generate_data()
sce <- loadData(cell_file)
sce <- transformLogScale(sce)

sce$Drug <- as.factor(sce$Drug)
sce$Drug <- relevel(sce$Drug, ref = "D1")
types <- c("AreaShape", "Intensity", "Texture")

sce_single <- predictLOO(
 sce,
 target = "Drug", group = "Patient",
 interest_level = "D7", reference_level = "D1",
 types = types,
 n_threads = 1
 )

aggregateYhat(sce_single, meta_vars = c("Patient", "Drug"))

}
