% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpaintr.R
\name{aggregateYhat}
\alias{aggregateYhat}
\title{Aggregate predicted leave-one-out probabilities over meta variables}
\usage{
aggregateYhat(
  sce,
  assay_type = "tfmfeatures",
  meta_vars = c("Patient", "Treatment", "Gender")
)
}
\arguments{
\item{sce}{\code{\link[SingleCellExperiment]{SingleCellExperiment}} object}

\item{assay_type}{A string specifying the assay}

\item{meta_vars}{a vector of variables from `colData`}
}
\value{
\code{\link[data.frame]{data.frame}}
}
\description{
Aggregate predicted leave-one-out probabilities over meta variables
}
