% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpaintr.R
\name{calculateStats}
\alias{calculateStats}
\title{Aggregate predicted leave-one-out probabilities over meta variables
over a list of SingleCellExperiment objects}
\usage{
calculateStats(
  sce,
  assay_type = "tfmfeatures",
  meta_vars = c("Patient", "Treatment"),
  target = "Treatment"
)
}
\arguments{
\item{sce}{A \code{\link[SingleCellExperiment]{SingleCellExperiment}} object}

\item{assay_type}{A string specifying the assay}

\item{meta_vars}{a vector of variables from `colData`}

\item{target}{Name of target variable for prediction}
}
\value{
\code{data.frame}
}
\description{
Aggregate predicted leave-one-out probabilities over meta variables
over a list of SingleCellExperiment objects
}
\examples{
set.seed(23)
cell_file <- generate_data()
sce <- loadData(cell_file)
sce <- transformLogScale(sce)

sce$Drug <- as.factor(sce$Drug)
sce$Drug <- relevel(sce$Drug, ref = "D1")
types <- c("AreaShape", "Intensity", "Texture")

sce_single <- predictLOO(
 sce,
 target = "Drug", group = "Patient",
 interest_level = "D7", reference_level = "D1",
 types = types,
 n_threads = 1
 )

calculateStats(sce_single, meta_vars = c("Patient", "Drug"), target = "Drug")

}
