% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpaintr.R
\name{calculateStats}
\alias{calculateStats}
\title{Aggregate predicted leave-one-out probabilities over meta variables
over a list of SingleCellExperiment objects}
\usage{
calculateStats(
  sce_list,
  assay_type = "tfmfeatures",
  meta_vars = c("Patient", "Treatment", "Gender"),
  target = "Treatment",
  panel = "Panel"
)
}
\arguments{
\item{assay_type}{A string specifying the assay}

\item{meta_vars}{a vector of variables from `colData`}

\item{target}{Name of target variable for prediction}

\item{panel}{Name of panel variable}

\item{sce}{\code{\link[SingleCellExperiment]{SingleCellExperiment}} object}
}
\value{
\code{\link[data.frame]{data.frame}}
}
\description{
Aggregate predicted leave-one-out probabilities over meta variables
over a list of SingleCellExperiment objects
}
