% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpaintr.R
\name{calculateStats}
\alias{calculateStats}
\title{Aggregate predicted leave-one-out probabilities over meta variables
over a list of SingleCellExperiment objects}
\usage{
calculateStats(
  sce_list,
  assay_type = "tfmfeatures",
  meta_vars = c("Patient", "Treatment"),
  target = "Treatment"
)
}
\arguments{
\item{sce_list}{A list of
\code{\link[SingleCellExperiment]{SingleCellExperiment}}
objects}

\item{assay_type}{A string specifying the assay}

\item{meta_vars}{a vector of variables from `colData`}

\item{target}{Name of target variable for prediction}
}
\value{
\code{\link[data.frame]{data.frame}}
}
\description{
Aggregate predicted leave-one-out probabilities over meta variables
over a list of SingleCellExperiment objects
}
