% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpaintr.R
\name{compute_y_hat}
\alias{compute_y_hat}
\title{Internal function to compute y_hat on a subset of the features}
\usage{
compute_y_hat(
  feature_name,
  sce_feature,
  starts,
  assay_type,
  target,
  interest_level,
  reference_level,
  group,
  weights,
  n_threads
)
}
\arguments{
\item{feature_name}{String with feature name}

\item{sce_feature}{\code{\link[SingleCellExperiment]{SingleCellExperiment}}}

\item{starts}{Starting string}

\item{assay_type}{A string specifying the assay}

\item{target}{Name of target variable for prediction}

\item{interest_level}{Factor interest level in `target` variable}

\item{reference_level}{Factor reference level in `target` variable}

\item{group}{Grouping variable for cross-validation, e.g., patient}

\item{weights}{Weights variable when features are aggregated}

\item{n_threads}{Number of parallel threads for fitting of models}
}
\value{
\code{\link[tibble]{tibble}} data frame
}
\description{
Internal function to compute y_hat on a subset of the features
}
