% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpaintr.R
\name{loadData}
\alias{loadData}
\title{Load cell painting data from file and convert to a SingleCellExperiment}
\usage{
loadData(cell_file)
}
\arguments{
\item{path}{features_file}
}
\value{
\code{\link[SingleCellExperiment]{SingleCellExperiment}} object
}
\description{
Load cell painting data from file and convert to a SingleCellExperiment
}
