% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpaintr.R
\name{plotAUC}
\alias{plotAUC}
\title{Plot AUC comparison}
\usage{
plotAUC(result_list)
}
\arguments{
\item{result_list}{\code{\link[list]{list}} of data frames from `fitModel`}
}
\value{
\code{\link[ggplot2]{ggplot2}} object
}
\description{
Plot AUC comparison
}
