% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpaintr.R
\name{plotAUC}
\alias{plotAUC}
\title{Plot AUC comparison}
\usage{
plotAUC(
  sce,
  assay_type = "tfmfeatures",
  meta_vars = c("Patient", "Treatment"),
  target = "Treatment",
  batch = NULL
)
}
\arguments{
\item{sce}{A single or list of
\code{\link[SingleCellExperiment]{SingleCellExperiment}} objects}

\item{assay_type}{A string specifying the assay}

\item{meta_vars}{a vector of variables from `colData`}

\item{target}{Name of target variable for prediction}

\item{batch}{Name of batch variable}
}
\value{
\code{\link[ggplot2]{ggplot2}} object
}
\description{
Plot AUC comparison
}
