% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpaintr.R
\name{plotCellsPerImage}
\alias{plotCellsPerImage}
\title{Plot number of cells per image}
\usage{
plotCellsPerImage(sce, bins = 100)
}
\arguments{
\item{sce}{\code{\link[SingleCellExperiment]{SingleCellExperiment}} object}

\item{bins}{Number of histogram bins}
}
\value{
\code{\link[ggplot2]{ggplot2}} object
}
\description{
Plot number of cells per image
}
