% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpaintr.R
\name{plotLOO}
\alias{plotLOO}
\title{Plot predicted leave-one-out probabilities}
\usage{
plotLOO(
  sce,
  assay_type = "tfmfeatures",
  meta_vars = c("Patient", "Treatment", "Gender"),
  target = "Treatment"
)
}
\arguments{
\item{sce}{\code{\link[SingleCellExperiment]{SingleCellExperiment}} object}

\item{assay_type}{A string specifying the assay}

\item{meta_vars}{a vector of variables from `colData`}

\item{target}{Name of target variable for prediction}
}
\value{
\code{\link[ggplot2]{ggplot2}} object
}
\description{
Plot predicted leave-one-out probabilities
}
