% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpaintr.R
\name{plotPCACor}
\alias{plotPCACor}
\title{Plot number of cells per image}
\usage{
plotPCACor(sce, filter_by = 1, top = 20, pcs = 1:5)
}
\arguments{
\item{sce}{\code{\link[SingleCellExperiment]{SingleCellExperiment}} object}

\item{filter_by}{PC to use for feature selection}

\item{top}{Number of top features to select}

\item{pcs}{Number of PCs to plot}
}
\value{
\code{\link[ggplot2]{ggplot2}} object
}
\description{
Plot number of cells per image
}
