% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpaintr.R
\name{plotROC}
\alias{plotROC}
\title{Plot ROC curves}
\usage{
plotROC(result)
}
\arguments{
\item{result}{\code{\link[tibble]{tibble}} data frame}
}
\value{
\code{\link[ggplot2]{ggplot2}} object
}
\description{
Plot ROC curves
}
