% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpaintr.R
\name{predictLOO}
\alias{predictLOO}
\title{Predict target from features}
\usage{
predictLOO(
  sce,
  target,
  group,
  interest_level,
  reference_level,
  types = NULL,
  channels = NULL,
  weights = NULL,
  n_threads = 1,
  assay_type = "tfmfeatures"
)
}
\arguments{
\item{sce}{\code{\link[SingleCellExperiment]{SingleCellExperiment}} object}

\item{target}{Name of target variable for prediction}

\item{group}{Grouping variable for cross-validation, e.g., patient}

\item{interest_level}{Factor interest level in `target` variable}

\item{reference_level}{Factor reference level in `target` variable}

\item{types}{Vector of strings of feature types}

\item{channels}{Vector of strings of feature channels}

\item{weights}{Weights variable when features are aggregated}

\item{n_threads}{Number of parallel threads for fitting of models}

\item{assay_type}{A string specifying the assay}
}
\value{
\code{\link[tibble]{tibble}} data frame
}
\description{
Predict target from features
}
\examples{
set.seed(23)
cell_file <- generate_data()
sce <- loadData(cell_file)
sce <- transformLogScale(sce)

sce$Drug <- as.factor(sce$Drug)
sce$Drug <- relevel(sce$Drug, ref = "D1")
types <- c("AreaShape", "Intensity", "Texture")

sce_single <- predictLOO(
    sce,
    target = "Drug", group = "Patient",
    interest_level = "D7", reference_level = "D1",
    types = types,
    n_threads = 1
)

}
