% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpaintr.R
\name{removeOutliers}
\alias{removeOutliers}
\title{Remove cells if not enough or too many in one image}
\usage{
removeOutliers(sce, min, max)
}
\arguments{
\item{sce}{\code{\link[SingleCellExperiment]{SingleCellExperiment}} object}

\item{min}{Remove cells below that number}

\item{max}{Remove cells above that number}
}
\value{
\code{\link[SingleCellExperiment]{SingleCellExperiment}} object
}
\description{
Remove cells if not enough or too many in one image
}
