% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpaintr.R
\name{transformLogScale}
\alias{transformLogScale}
\title{Filter low variance features}
\usage{
transformLogScale(sce, robust = FALSE)
}
\arguments{
\item{sce}{\code{\link[SingleCellExperiment]{SingleCellExperiment}} object}

\item{robust}{If true robust z-score, otherwise standard z-score}
}
\value{
\code{\link[SingleCellExperiment]{SingleCellExperiment}} object
}
\description{
Filter low variance features
}
