% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpaintr.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Plot predicted leave-one-out probabilities}
\usage{
volcanoPlot(
  sce,
  assay_type = "tfmfeatures",
  meta_vars = c("Patient", "Treatment"),
  target = "Treatment",
  p_cutoff = NULL,
  fc_cutoff = 1
)
}
\arguments{
\item{sce}{A \code{\link[SingleCellExperiment]{SingleCellExperiment}} object}

\item{assay_type}{A string specifying the assay}

\item{meta_vars}{a vector of variables from `colData`}

\item{target}{Name of target variable for prediction}

\item{p_cutoff}{Cut-off for statistical significance. A horizontal line
will be drawn at -log10(p_cutoff).}

\item{fc_cutoff}{Cut-off for absolute log2 fold-change. A vertical lines
will be drawn at fc_cutoff.}
}
\value{
\code{\link[ggplot2]{ggplot2}} object
}
\description{
Plot predicted leave-one-out probabilities
}
\examples{
set.seed(23)
cell_file <- generate_data()
sce <- loadData(cell_file)
sce <- transformLogScale(sce)

sce$Drug <- as.factor(sce$Drug)
sce$Drug <- relevel(sce$Drug, ref = "D1")
types <- c("AreaShape", "Intensity", "Texture")

sce_single <- predictLOO(
    sce,
    target = "Drug", group = "Patient",
    interest_level = "D7", reference_level = "D1",
    types = types,
    n_threads = 1
)

volcanoPlot(sce_single,
    meta_vars = c("Patient", "Drug"), target = "Drug",
    p_cutoff = 0.05, fc_cutoff = 0.5
)

}
