% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpaintr.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Plot predicted leave-one-out probabilities}
\usage{
volcanoPlot(
  stats,
  p_cutoff = 1e-04,
  fc_cutoff = 0.5,
  color = "Panel",
  label = "Target"
)
}
\arguments{
\item{stats}{a \code{\link[data.frame]{data.frame}} from `calculateStats`}

\item{p_cutoff}{Cut-off for statistical significance. A horizontal line will
be drawn at -log10(p_cutoff).}

\item{fc_cutoff}{Cut-off for absolute log2 fold-change. A vertical lines
will be drawn at fc_cutoff.}

\item{color}{Name of color variable}

\item{label}{Name of label variable}
}
\value{
\code{\link[ggplot2]{ggplot2}} object
}
\description{
Plot predicted leave-one-out probabilities
}
