% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpaintr.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Plot predicted leave-one-out probabilities}
\usage{
volcanoPlot(
  sce_list,
  assay_type = "tfmfeatures",
  meta_vars = c("Patient", "Treatment", "Gender"),
  target = "Treatment",
  p_cutoff = 0.01/100,
  fc_cutoff = 0.3
)
}
\arguments{
\item{sce_list}{A list of
\code{\link[SingleCellExperiment]{SingleCellExperiment}}
objects}

\item{assay_type}{A string specifying the assay}

\item{meta_vars}{a vector of variables from `colData`}

\item{target}{Name of target variable for prediction}

\item{p_cutoff}{Cut-off for statistical significance. A horizontal line will
be drawn at -log10(p_cutoff).}

\item{fc_cutoff}{Cut-off for absolute log2 fold-change. A vertical lines
will be drawn at fc_cutoff.}
}
\value{
\code{\link[ggplot2]{ggplot2}} object
}
\description{
Plot predicted leave-one-out probabilities
}
